StbHardware.py: long() is deprecated and merged with int
File "/usr/lib/enigma2/python/Tools/StbHardware.py", line 11, in getFPVersion
    ret = long(open("/proc/stb/fp/version", "r").read())
NameError: name 'long' is not defined

+++ lib/python/Tools/StbHardware.py
@@ -8,7 +8,7 @@ from time import time, localtime, gmtime
 def getFPVersion():
 	ret = None
 	try:
-		ret = long(open("/proc/stb/fp/version", "r").read())
+		ret = open("/proc/stb/fp/version", "r").read()
 	except IOError:
 		try:
 			fp = open("/dev/dbox/fp0")
@@ -58,7 +58,7 @@ def setRTCtime(wutime):
 def getFPWakeuptime():
 	ret = 0
 	try:
-		ret = long(open("/proc/stb/fp/wakeup_time", "r").read())
+		ret = open("/proc/stb/fp/wakeup_time", "r").read()
 	except IOError:
 		try:
 			fp = open("/dev/dbox/fp0")
